added batchTrace functionality
If the model sets the parameter batch-trace to 't', a .tracedat file will be created with the same name as the .dat file. This file will contain the trace of the model.
The batch trace outputs the operators that have fired, the retrievals performed, the actions performed, and the input to the visual buffer, plus their time stamps.

Be careful for using this on really large batch runs because the file can get very big - maybe we should implement a warning for this?